refactored, tests need fixing/writing
